iommu/arm-smmu: Add support for MMU40x/500 clocks
The MMU400x/500 is the implementation of the SMMUv2
arch specification. It is split in to two blocks
TBU, TCU. TBU caches the page table, instantiated
for each master locally, clocked by the TBUn_clk.
TCU manages the address translation with PTW and has
the programming interface as well, clocked using the
TCU_CLK. The TBU can also be sharing the same clock
domain as TCU, in which case both are clocked using
the TCU_CLK.

This defines the clock bindings for the same and adds
the clock names to comaptible data.

Signed-off-by: Sricharan R <sricharan@codeaurora.org>
[vivek: add clock name data for MMU500, removed separate clock init,
        enable and disable calls for MMU500]
Signed-off-by: Vivek Gautam <vivek.gautam@codeaurora.org>