soc: qcom: rpmh: Fix TCS offset
An available TCS is picked and the tcs group-local index is calculated,
a response object is created and a lookup for the tcs group is done
based on the previously acquired index, but now in global number space.

The result is that the response object is associated with the wrong tcs
group, which isn't locked and we're hitting the lock assert in
setup_response().

Signed-off-by: Bjorn Andersson <bjorn.andersson@linaro.org>