iommu/arm-smmu: Error out only if not enough context interrupts
Currently we check if the number of context banks are unequal to
num_context_interrupts. However, there are booloaders such as, one
on sdm845 that can mask few context banks from the kernel.
So, although the hardware definition in device tree would mention
the correct number of context interrupts, this number can be
greater than the number of context banks visible to smmu in kernel.
We should therefore error out only when the number of context banks
is greater than the available number of context interrupts.

Signed-off-by: Vivek Gautam <vivek.gautam@codeaurora.org>
Suggested-by: Tomasz Figa <tfiga@chromium.org>
Cc: Robin Murphy <robin.murphy@arm.com>
Cc: Will Deacon <will.deacon@arm.com>